Compute Median Prices per Administrative Level

Usage

admin_level_medians(df, admin_level = "municipality", precision = 3)

Arguments

df

Monthly JMMI clean dataset

admin_level

an administrative level in Libya, can be one of "municipality", "district", "region", or "overall"

precision

The number of decimal places to keep when rounding results

Value

A dataframe with median prices for each item in the MEB per Administrative Level

Examples

admin_level_medians(jmmi_2022_feb, "region")
#> # A tibble: 135 × 3
#>    q_region         item                             median_item_price
#>    <chr>            <chr>                                        <dbl>
#>  1 East (Cyrenaica) cooking_fuel_price_per_11kg                  16.9 
#>  2 East (Cyrenaica) q_Ibuprofen_price_per_20                      6   
#>  3 East (Cyrenaica) q_amoxicillin_price_per_21                   10   
#>  4 East (Cyrenaica) q_beans_price_per_400g                        2.5 
#>  5 East (Cyrenaica) q_bread_price_per_5medium_pieces              1.25
#>  6 East (Cyrenaica) q_btea_price_per_250g                         6   
#>  7 East (Cyrenaica) q_chicken_price_per_kilo                     11   
#>  8 East (Cyrenaica) q_chickpeas_price_per_400g                    2.5 
#>  9 East (Cyrenaica) q_cmilk_price_per_200ml                       2.5 
#> 10 East (Cyrenaica) q_couscous_price_per_kilo                     5   
#> # … with 125 more rows
admin_level_medians(jmmi_2022_feb, "district")
#> # A tibble: 990 × 3
#>    q_district         item                             median_item_price
#>    <chr>              <chr>                                        <dbl>
#>  1 Al Jabal Al Akhdar cooking_fuel_price_per_11kg                  27.5 
#>  2 Al Jabal Al Akhdar q_Ibuprofen_price_per_20                     10   
#>  3 Al Jabal Al Akhdar q_amoxicillin_price_per_21                    6.3 
#>  4 Al Jabal Al Akhdar q_beans_price_per_400g                        2.25
#>  5 Al Jabal Al Akhdar q_bread_price_per_5medium_pieces              1.25
#>  6 Al Jabal Al Akhdar q_btea_price_per_250g                         5   
#>  7 Al Jabal Al Akhdar q_chicken_price_per_kilo                     10.8 
#>  8 Al Jabal Al Akhdar q_chickpeas_price_per_400g                    2.2 
#>  9 Al Jabal Al Akhdar q_cmilk_price_per_200ml                       3.12
#> 10 Al Jabal Al Akhdar q_couscous_price_per_kilo                     5   
#> # … with 980 more rows
admin_level_medians(jmmi_2022_feb, "municipality")
#> # A tibble: 1,755 × 3
#>    q_municipality item                             median_item_price
#>    <chr>          <chr>                                        <dbl>
#>  1 Abusliem       cooking_fuel_price_per_11kg                   5.75
#>  2 Abusliem       q_Ibuprofen_price_per_20                     10   
#>  3 Abusliem       q_amoxicillin_price_per_21                   22   
#>  4 Abusliem       q_beans_price_per_400g                        2.75
#>  5 Abusliem       q_bread_price_per_5medium_pieces              1.12
#>  6 Abusliem       q_btea_price_per_250g                         6.25
#>  7 Abusliem       q_chicken_price_per_kilo                      8.75
#>  8 Abusliem       q_chickpeas_price_per_400g                    2.25
#>  9 Abusliem       q_cmilk_price_per_200ml                       2.88
#> 10 Abusliem       q_couscous_price_per_kilo                     3.5 
#> # … with 1,745 more rows